requirements.txt
openai==1.51.2
python-dotenv==1.1.1